backend/spootify/requirements.txt
apipkg==3.0.2
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.38.18
botocore==1.38.18
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
cryptography==45.0.2
daphne==4.2.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
dnspython==2.7.0
git-filter-repo==2.47.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
msgpack==1.1.0
psycopg==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.3
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
redis==6.1.0
s3transfer==0.12.0
service-identity==24.2.0
setuptools==76.1.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.4.0
zope.interface==7.2
